requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
passlib==1.7.4
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
python-dotenv==1.0.1
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
